# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "0.1"
    },
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users list",
                "description": "Get list of users",
                "operationId": "getListOfUsers",
                "responses": {
                    "200": {
                        "description": "Get list of users.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "surname": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            },
                                                            "path": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add user",
                "description": "Add user",
                "operationId": "e62c685cc38d5057ff0eee6e66bc7761",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "surname"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Fox"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 70
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Anna"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "anna@gmail.com"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Rutkowska"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "44ee65e3907e090f44bf2ebdc07ecbf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 70
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Fox"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 70
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Anna"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "anna@gmail.com"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Rutkowska"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "807de0d0c75fe6828b0d0e70c26fcda5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/user": {
            "get": {
                "tags": [
                    "UsersCQRS"
                ],
                "summary": "Users list",
                "description": "Get list of users CQRS",
                "operationId": "getListOfUsersCQRS",
                "responses": {
                    "200": {
                        "description": "Get list of users.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "surname": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            },
                                                            "path": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UsersCQRS"
                ],
                "summary": "Add user",
                "description": "Add user",
                "operationId": "cdba9bb1d60f3a456388b1d4dcb3e122",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "surname"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Fox"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 70
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Anna"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "anna@gmail.com"
                                                        },
                                                        "surname": {
                                                            "type": "string",
                                                            "example": "Rutkowska"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}